+++ src/uint.rs
@@ -10,8 +10,6 @@
 mod concat;
 #[macro_use]
 mod split;
-#[macro_use]
-mod convert;
 
 mod add;
 mod add_mod;
@@ -45,6 +43,9 @@ mod array;
 
 #[cfg(feature = "rand_core")]
 mod rand;
+
+#[cfg(feature = "extra-trait-impls")]
+mod extra_trait_impls;
 
 use crate::{Bounded, Concat, Encoding, Integer, Limb, Split, Word, Zero};
 use core::fmt;
@@ -409,246 +410,6 @@ impl_split! {
     (U4096, 4096),
     (U6144, 6144),
     (U8192, 8192)
-}
-
-impl_convert! {
-    (U64, 64),
-    (
-    (U128, 128),
-    (U256, 256),
-    (U384, 384),
-    (U512, 512),
-    (U640, 640),
-    (U768, 768),
-    (U896, 896),
-    (U1024, 1024),
-    (U1280, 1280),
-    (U1536, 1536),
-    (U1792, 1792),
-    (U2048, 2048),
-    (U3072, 3072),
-    (U3584, 3584),
-    (U4096, 4096),
-    (U6144, 6144),
-    (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U128, 128),
-    (
-        (U256, 256),
-        (U384, 384),
-        (U512, 512),
-        (U640, 640),
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U256, 256),
-    (
-        (U384, 384),
-        (U512, 512),
-        (U640, 640),
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U384, 384),
-    (
-        (U512, 512),
-        (U640, 640),
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-impl_convert! {
-    (U512, 512),
-    (
-        (U640, 640),
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-impl_convert! {
-    (U640, 640),
-    (
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-impl_convert! {
-    (U768, 768),
-    (
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U896, 896),
-    (
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U1024, 1024),
-    (
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U1280, 1280),
-    (
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U1536, 1536),
-    (
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U2048, 2048),
-    (
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U3072, 3072),
-    (
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U3584, 3584),
-    (
-        (U4096, 4096),
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U4096, 4096),
-    (
-        (U6144, 6144),
-        (U8192, 8192)
-    )
-}
-
-impl_convert! {
-    (U6144, 6144),
-    (
-        (U8192, 8192)
-    )
 }
 
 #[cfg(test)]
